This commit fixes several errors and the blank pages
What little exposure to Ruby has convinced me that it is not
ready for production software due to the snarling dependencies
and incompatible versions, and of course wierd errors.

This is no exception.  Git hub uses a stable version of Jekyll,
while the default available to install uses a more recent version.
The theme was locked to an older version... causing headaches.

The documentation is wrong, and this wasted thirty minutes of
my time to resolve by downgrading my locally installed version
of Jekyll, and figuring out that the theme layout conventions
aren't standard throughout themes.  What a way to waste a
Saturday morning.